requirements.txt
amqp==2.6.1
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
billiard==3.6.4.0
cairocffi==1.3.0
CairoSVG==2.5.2
celery==4.4.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
cssselect2==0.6.0
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==3.2.9
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-mptt-graph==0.3
django-redis-cache==2.1.0
django-rest-swagger==2.2.0
django-rq==2.1.0
django-translation-manager==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docxcompose==1.3.4
docxtpl==0.16.0
drf-yasg==1.20.0
environs==9.5.0
et-xmlfile==1.1.0
flower==0.9.7
generics==3.9.0
gunicorn==20.0.0
html5lib==1.1
humanize==4.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.2
kombu==4.6.11
lxml==4.8.0
MarkupSafe==2.0.1
marshmallow==3.15.0
model-bakery==1.4.0
openapi-codec==1.3.2
openpyxl==3.0.10
packaging==21.2
pendulum==2.1.2
Pillow==8.4.0
pluggy==1.0.0
polib==1.1.1
prometheus-client==0.8.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.10.0
pyTelegramBotAPI==3.7.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.20.0
pytz==2021.3
pytzdata==2020.1
qrcode==7.3.1
redis==3.5.3
requests==2.26.0
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
supervisor==4.2.2
Telethon==1.25.0
tinycss2==1.1.0
tomli==2.0.1
tornado==6.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
WeasyPrint==52.5
webencodings==0.5.1
xlrd==2.0.1
